safely delete temp files
Previously, our test assumed that any file prefixed with
"file-" was a file that we created as part of the test.
This is not a safe assumption since we may unintentionally
delete other files with that prefix that weren't created
by our test.

In order to fix this issue, this commit changes which
files we remove as part of the cleanup. We now reference
the exact files that were created for the test instead
of assuming that any file prefixed by "file-" was.

Note that the test cleanup yields undefined behavior
if the function being tested fails. It's possible that
the test fails with files left over or extra files
that were deleted but shouldn't have been.